clean up insar addition to hyp3-gamma
