+++ tests/test_entrypoints.py
@@ -29,3 +29,31 @@ def test_hyp3_sentinel(script_runner):
 def test_smooth_dem_tiles(script_runner):
     ret = script_runner.run('smooth_dem_tiles.py', '-h')
     assert ret.success
+
+
+def test_insar_passthrough(script_runner):
+    ret = script_runner.run(
+        'hyp3_gamma', '++process', 'insar', '--help')
+    assert ret.success
+    assert 'insar' in ret.stdout
+    assert '--include-los-displacement' in ret.stdout
+
+
+def test_insar_gamma(script_runner):
+    default_help = script_runner.run('insar', '-h')
+    assert default_help.success
+
+
+def test_ifm_sentinel(script_runner):
+    ret = script_runner.run('ifm_sentinel.py', '-h')
+    assert ret.success
+
+
+def test_interf_pwr_s1_lt_tops_proc(script_runner):
+    ret = script_runner.run('interf_pwr_s1_lt_tops_proc.py', '-h')
+    assert ret.success
+
+
+def test_unwrapping_geocoding(script_runner):
+    ret = script_runner.run('unwrapping_geocoding.py', '-h')
+    assert ret.success
